requirements.txt
# requirements.txt
# Instalar: pip install -r requirements.txt

# LLM
ollama>=0.4.0
groq>=0.13.0

# Banco de dados
sqlalchemy>=2.0.0
pymysql>=1.1.0
duckdb>=1.0.0

# Analise de dados
pandas>=2.2.0
polars>=1.0.0
openpyxl>=3.1.0

# Interface
streamlit>=1.32.0

# Ambiente
python-dotenv>=1.0.0

# Notebooks
jupyter>=1.1.0
ipykernel>=6.29.0
